Update int ranges
