+++ crates/polars-plan/src/dsl/function_expr/range/int_range.rs
@@ -1,8 +1,10 @@
+use std::iter::zip;
+
 use polars_core::prelude::*;
 use polars_core::series::{IsSorted, Series};
 use polars_core::with_match_physical_integer_polars_type;
 
-use super::utils::{broadcast_scalar_inputs, ensure_range_bounds_contain_exactly_one_value};
+use super::utils::{ensure_range_bounds_contain_exactly_one_value, ranges_impl_broadcast};
 
 pub(super) fn int_range(s: &[Series], step: i64, dtype: DataType) -> PolarsResult<Series> {
     let mut start = &s[0];
@@ -72,8 +74,6 @@ pub(super) fn int_ranges(s: &[Series], step: i64) -> PolarsResult<Series> {
     let start = &s[0].rechunk();
     let end = &s[1].rechunk();
 
-    let output_name = "int_range";
-
     let mut start = start.cast(&DataType::Int64)?;
     let mut end = end.cast(&DataType::Int64)?;
 
@@ -82,47 +82,61 @@ pub(super) fn int_ranges(s: &[Series], step: i64) -> PolarsResult<Series> {
     let start = start.i64()?;
     let end = end.i64()?;
 
-    let start = start.downcast_iter().next().unwrap();
-    let end = end.downcast_iter().next().unwrap();
-
     // First do a pass to determine the required value capacity.
-    let mut values_capacity = 0;
-    for (opt_start, opt_end) in start.into_iter().zip(end) {
+    let mut capacity = 0;
+    let mut increment_capacity = |start: i64, end: i64| {
+        if step == 1 {
+            capacity += (end - start).unsigned_abs() as usize;
+        } else {
+            capacity += (((end - start).unsigned_abs() / step.unsigned_abs()) + 1) as usize;
+        }
+    };
+    for (opt_start, opt_end) in zip(start, end) {
         if let (Some(start_v), Some(end_v)) = (opt_start, opt_end) {
-            if step == 1 {
-                values_capacity += (end_v - start_v).unsigned_abs() as usize;
-            } else {
-                values_capacity +=
-                    (((end_v - start_v).unsigned_abs() / step.unsigned_abs()) + 1) as usize;
-            }
+            increment_capacity(start_v, end_v)
         }
     }
 
     let mut builder = ListPrimitiveChunkedBuilder::<Int64Type>::new(
-        output_name,
+        "int_range",
         start.len(),
-        values_capacity,
+        capacity,
         DataType::Int64,
     );
 
-    for (opt_start, opt_end) in start.into_iter().zip(end) {
-        match (opt_start, opt_end) {
-            (Some(&start_v), Some(&end_v)) => match step {
-                1 => {
-                    builder.append_iter_values(start_v..end_v);
-                },
-                2.. => {
-                    builder.append_iter_values((start_v..end_v).step_by(step as usize));
-                },
-                _ => builder.append_iter_values(
-                    (end_v..start_v)
-                        .step_by(step.unsigned_abs() as usize)
-                        .map(|x| start_v - (x - end_v)),
-                ),
-            },
-            _ => builder.append_null(),
-        }
-    }
+    let range_impl = |start, end, builder: &mut ListPrimitiveChunkedBuilder<Int64Type>| {
+        match step {
+            1 => builder.append_iter_values(start..end),
+            2.. => builder.append_iter_values((start..end).step_by(step as usize)),
+            _ => builder.append_iter_values(
+                (end..start)
+                    .step_by(step.unsigned_abs() as usize)
+                    .map(|x| start - (x - end)),
+            ),
+        };
+        Ok(())
+    };
 
-    Ok(builder.finish().into_series())
+    ranges_impl_broadcast(start, end, range_impl, &mut builder)
+}
+
+fn broadcast_scalar_inputs(start: Series, end: Series) -> PolarsResult<(Series, Series)> {
+    match (start.len(), end.len()) {
+        (len1, len2) if len1 == len2 => Ok((start, end)),
+        (1, len2) => {
+            let start_matched = start.new_from_index(0, len2);
+            Ok((start_matched, end))
+        },
+        (len1, 1) => {
+            let end_matched = end.new_from_index(0, len1);
+            Ok((start, end_matched))
+        },
+        (len1, len2) => {
+            polars_bail!(
+                ComputeError:
+                "lengths of `start` ({}) and `end` ({}) do not match",
+                len1, len2
+            )
+        },
+    }
 }
